refactor(chi): migrate rest microservices to chi
