+++ cmd/chat/main.go
@@ -1,15 +1,10 @@
 package main
 
 import (
-	"fmt"
 	"log"
-	"net/http"
 
 	"github.com/hexley21/fixup/pkg/config"
-	"github.com/hexley21/fixup/pkg/logger"
-	"github.com/hexley21/fixup/internal/common/rest"
-	"github.com/labstack/echo/v4"
-	"github.com/labstack/echo/v4/middleware"
+	"github.com/hexley21/fixup/pkg/logger/zap_logger"
 )
 
 func main() {
@@ -18,26 +13,19 @@ func main() {
 		log.Fatalf("could not load config: %v\n", err)
 	}
 
-	e := echo.New()
+	zapLogger := zap_logger.New(cfg.Logging, cfg.Server.IsProd)
+	zapLogger.Debug("")
+	// playgroundValidator := playground_validator.New()
 
-	e.Logger = logger.NewZapLogger(cfg.Logging, cfg.Server.IsProd)
+	// pgPool, err := postgres.NewPool(&cfg.Postgres)
+	// if err != nil {
+	// 	zapLogger.Fatal(err)
+	// }
 
-	e.Use(middleware.Logger())
-	e.Use(middleware.Recover())
+	// snowflakeNode, err := snowflake.NewNode(cfg.Server.InstanceId)
+	// if err != nil {
+	// 	zapLogger.Fatal(err)
+	// }
 
-	e.GET("/", func(c echo.Context) error {
-		return c.String(http.StatusOK, "Hello World")
-	})
-
-	e.HTTPErrorHandler = func(err error, c echo.Context) {
-		if apiErr, ok := err.(*rest.ErrorResponse); ok {
-			c.JSON(apiErr.Status, apiErr)
-			c.Logger().Error(err)
-			return
-		}
-		c.Logger().Error(err)
-		c.JSON(http.StatusInternalServerError, rest.NewInternalServerError(err))
-	}
-
-	e.Logger.Fatal(e.Start(fmt.Sprintf(":%d", cfg.Server.HttpPort)))
+	
 }
