+++ config/order.config.yml
@@ -1,8 +1,13 @@
 server:
-    http_port: 80
     instance_id: 1
     shutdown_timeout: 10s
+    email: order@fixup.com
 
+http:
+    port: 80
+    idle_timeout: 60s
+    read_timeout: 10s
+    write_timeout: 30s
 
 postgres:
     port: 5432
