+++ config/user.config.yml
@@ -1,6 +1,7 @@
 server:
     instance_id: 1
     shutdown_timeout: 10s
+    email: auth@fixup.com
 
 http:
     port: 80
@@ -24,7 +25,7 @@ redis:
     db_name: 0
     dial_timeout: 30s
 
-aws: 
+aws:
     awscfg:
         region: eu-north-1
     s3:
@@ -33,7 +34,6 @@ aws:
     cdn:
         url_fmt: https://d20eri1dy5h30b.cloudfront.net/%s
         expiry: 24h
-
 
 jwt:
     access_ttl: 2h
