+++ pkg/validator/playground_validator/playground_validator.go
@@ -1,4 +1,4 @@
-package plaground_validator
+package playground_validator
 
 import (
 	"regexp"
